+++ modules/Students/includes/Other_Fields.inc.php
@@ -23,7 +23,6 @@ foreach ( (array) $fields_RET as $field )
 			break;
 
 		case 'autos':
-		case 'edits':
 
 			echo _makeAutoSelectInput( 'CUSTOM_' . $field['ID'], $field['TITLE'], $request );